fix(): refactor generators and transformers
