+++ bricks/vb-experiment/src/data-providers/utils/constants.ts
@@ -1,0 +1,10 @@
+export const REVERSED_ROUTE_KEYS: string[] = [
+  "context",
+  "exact",
+  "path",
+  "slot",
+];
+export const REVERSED_BRICK_KEYS: string[] = ["id", "ref", "slot"];
+export const REVERSED_TEMPLATE_KEYS: string[] = ["proxy", "state", "name"];
+
+export const IMPORT_DEFAULT_PREFIX = "default:";
