+++ bricks/vb-experiment/src/data-providers/utils/transformers/transformJsxChild.ts
@@ -1,0 +1,8 @@
+import * as t from "@babel/types";
+
+export function transformJsxChild(child: t.JSXElement | t.StringLiteral) {
+  if (child.type === "StringLiteral") {
+    return t.jsxExpressionContainer(child);
+  }
+  return child;
+}
